requirements.txt
# ============================================================
# CAAS — ChiangMai Air Quality Alert System
# Requirements File
# Usage: pip install -r requirements.txt
# ============================================================

# --- Data & Processing ---
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.2           # Read Excel files
python-dotenv==1.0.1       # Load .env variables

# --- API & Data Fetching ---
openmeteo-requests==1.3.0  # Open-Meteo weather API
requests==2.32.3
requests-cache==1.2.1      # Cache API responses
retry-requests==2.0.0      # Retry on failure

# --- Machine Learning ---
scikit-learn==1.5.1
xgboost==2.1.1

# --- Deep Learning (LSTM) ---
tensorflow==2.17.0

# --- MLOps & Experiment Tracking ---
mlflow==2.15.1
setuptools==80.9.0        # Keep pkg_resources available for mlflow 2.15.x

# --- Model Serving ---
fastapi==0.112.0
uvicorn[standard]==0.30.6
pydantic>=2.0

# --- Dashboard ---
streamlit==1.37.1
altair>=5.0               # Charts in Streamlit

# --- Monitoring & Drift Detection ---
evidently==0.4.33
scipy>=1.11               # KS test in monitoring script
joblib>=1.3

# --- Cloud (AWS) ---
boto3==1.35.0               # AWS S3, Lambda

# --- Testing ---
pytest==8.3.2
pytest-cov==5.0.0

# --- Notebooks (optional) ---
jupyter==1.0.0
